requirements.txt
aiogram~=3.18.0
pip~=23.2.1
attrs~=25.1.0
pytz~=2025.1
idna~=3.10
multidict~=6.1.0
propcache~=0.3.0
typing_extensions~=4.12.2
pydantic~=2.10.6
aiofiles~=24.1.0
magic-filter~=1.0.12
certifi~=2025.1.31
aiohttp~=3.11.12
yarl~=1.18.3
aiosignal~=1.3.2
frozenlist~=1.5.0
aiohappyeyeballs~=2.4.6
pydantic_core~=2.27.2
annotated-types~=0.7.0
python-decouple~=3.8